// client/src/interfaces/UserFieldErrors.ts
export interface UserFieldErrors {
    first_name?: string[];
    middle_name?: string[];
    last_name?: string[];
    suffix_name?: string[];
    gender?: string[];
    gender_id?: string[];
    birth_date?: string[];
    username?: string[];
    password?: string[];
    password_confirmation?: string[];
}